[repo] Move remaining samlang-core code into compiler crate
