+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
   "crates/samlang-cli",
   "crates/samlang-collections",
   "crates/samlang-compiler",
-  "crates/samlang-core",
   "crates/samlang-errors",
   "crates/samlang-heap",
   "crates/samlang-optimization",
