+++ crates/samlang-cli/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 pretty_assertions = "1.3.0"
 samlang-ast = { path = "../samlang-ast" }
-samlang-core = { path = "../samlang-core" }
+samlang-compiler = { path = "../samlang-compiler" }
 samlang-errors = { path = "../samlang-errors" }
 samlang-heap = { path = "../samlang-heap" }
 samlang-parser = { path = "../samlang-parser" }
