+++ crates/samlang-compiler/Cargo.toml
@@ -11,7 +11,9 @@ samlang-checker = { path = "../samlang-checker" }
 samlang-collections = { path = "../samlang-collections" }
 samlang-errors = { path = "../samlang-errors" }
 samlang-heap = { path = "../samlang-heap" }
+samlang-optimization = { path = "../samlang-optimization" }
 samlang-parser = { path = "../samlang-parser" }
+samlang-profiling = { path = "../samlang-profiling" }
 wat = "1.0.66"
 
 [dev-dependencies]
