+++ crates/samlang-compiler/src/loader.js
