--- crates/samlang-core/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "samlang-core"
-version = "0.9.1"
-edition = "2021"
-
-[dependencies]
-enum-as-inner = "0.5.1"
-itertools = "0.12.0"
-samlang-ast = { path = "../samlang-ast" }
-samlang-checker = { path = "../samlang-checker" }
-samlang-compiler = { path = "../samlang-compiler" }
-samlang-errors = { path = "../samlang-errors" }
-samlang-heap = { path = "../samlang-heap" }
-samlang-optimization = { path = "../samlang-optimization" }
-samlang-parser = { path = "../samlang-parser" }
-samlang-profiling = { path = "../samlang-profiling" }
